refactor: move reducers to own folder
Old thinking was that reducers should be closest to what calls it, but I realize that in this application reducers are used amongst various applications, and thus should be in their own folder
