+++ src/app/store.js
@@ -1,15 +1,15 @@
 import { configureStore } from 'redux-starter-kit';
 
-import rootReducer from './rootReducer';
+import rootReducer from 'reducers/rootReducer';
 
 const store = configureStore({
   reducer: rootReducer
 });
 
 if (process.env.NODE_ENV === 'development' && module.hot) {
-  module.hot.accept('./rootReducer', () => {
+  module.hot.accept('reducers/rootReducer', () => {
     // eslint-disable-next-line global-require
-    const newRootReducer = require('./rootReducer').default;
+    const newRootReducer = require('reducers/rootReducer').default;
     store.replaceReducer(newRootReducer);
   });
 }
